requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiostream==0.5.2
alert==0.0.2
annotated-types==0.5.0
antiorm==1.2.1
anyio==3.7.1
apturl==0.5.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.2.0
ChatterBot==1.0.0
chatterbot-corpus==1.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.17
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
contextlib2==21.6.0
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.5.14
db==0.1.1
db-sqlite3==0.0.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
embedchain==0.1.17
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.99.1
fasteners==0.14.1
Flask==2.3.3
flask-ngrok==0.0.25
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.2
google-auth==2.23.4
googleapis-common-protos==1.61.0
greenlet==2.0.2
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.0
httpx==0.25.1
humanfriendly==10.0
idna==3.3
importlib-metadata==6.8.0
importlib-resources==5.2.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.0.1
jsonpatch==1.33
jsonpointer==2.4
keyring==23.5.0
kubernetes==28.1.0
langchain==0.0.336
langsmith==0.0.65
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-index==0.8.69.post2
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
marshmallow==3.20.1
mathparse==0.1.2
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
netifaces==0.11.0
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.15.1
openai==1.3.3
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
overrides==7.4.0
packaging==20.9
pandas==2.1.3
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
Pillow==9.0.1
Pint==0.17
posthog==3.0.2
prompt-toolkit==3.0.41
protobuf==4.25.1
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.17.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==3.11.4
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==3.15.4
PyPika==0.48.9
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytube==15.0.0
pytz==2021.1
pyxdg==0.27
PyYAML==6.0.1
regex==2023.8.8
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schema==0.7.5
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.66.1
traitlets==5.13.0
twilio==8.7.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.23.2
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.20.0
wcwidth==0.2.10
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.16.0
xdg==5
xkit==0.0.0
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.4.1
